refactor: Move creation logic out of Review component
