+++ src/components/new-safe/create/logic/index.test.ts
@@ -7,7 +7,6 @@ import {
   relaySafeCreation,
   handleSafeCreationError,
 } from '@/components/new-safe/create/logic/index'
-
 import { type ErrorCode } from 'ethers'
 import { EthersTxReplacedReason } from '@/utils/ethers-utils'
 import { SafeCreationStatus } from '@/components/new-safe/create/steps/StatusStep/useSafeCreation'
@@ -58,237 +57,235 @@ jest.mock('@safe-global/protocol-kit', () => {
   }
 })
 
-describe('create logic', () => {
-  describe('checkSafeCreationTx', () => {
-    let waitForTxSpy = jest.spyOn(provider, 'waitForTransaction')
+describe('checkSafeCreationTx', () => {
+  let waitForTxSpy = jest.spyOn(provider, 'waitForTransaction')
 
-    beforeEach(() => {
-      jest.resetAllMocks()
+  beforeEach(() => {
+    jest.resetAllMocks()
 
-      jest.spyOn(web3, 'getWeb3ReadOnly').mockImplementation(() => provider)
+    jest.spyOn(web3, 'getWeb3ReadOnly').mockImplementation(() => provider)
 
-      waitForTxSpy = jest.spyOn(provider, 'waitForTransaction')
-      jest.spyOn(provider, 'getBlockNumber').mockReturnValue(Promise.resolve(4))
-      jest.spyOn(provider, 'getTransaction').mockReturnValue(Promise.resolve(mockTransaction as TransactionResponse))
-    })
+    waitForTxSpy = jest.spyOn(provider, 'waitForTransaction')
+    jest.spyOn(provider, 'getBlockNumber').mockReturnValue(Promise.resolve(4))
+    jest.spyOn(provider, 'getTransaction').mockReturnValue(Promise.resolve(mockTransaction as TransactionResponse))
+  })
 
-    it('returns SUCCESS if promise was resolved', async () => {
-      const receipt = {
-        status: 1,
-      } as TransactionReceipt
+  it('returns SUCCESS if promise was resolved', async () => {
+    const receipt = {
+      status: 1,
+    } as TransactionReceipt
 
-      waitForTxSpy.mockImplementationOnce(() => Promise.resolve(receipt))
+    waitForTxSpy.mockImplementationOnce(() => Promise.resolve(receipt))
 
-      const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+    const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
 
-      expect(result).toBe(SafeCreationStatus.SUCCESS)
-    })
+    expect(result).toBe(SafeCreationStatus.SUCCESS)
+  })
 
-    it('returns REVERTED if transaction was reverted', async () => {
-      const receipt = {
+  it('returns REVERTED if transaction was reverted', async () => {
+    const receipt = {
+      status: 0,
+    } as TransactionReceipt
+
+    waitForTxSpy.mockImplementationOnce(() => Promise.resolve(receipt))
+
+    const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+
+    expect(result).toBe(SafeCreationStatus.REVERTED)
+  })
+
+  it('returns TIMEOUT if transaction could not be found within the timeout limit', async () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TIMEOUT' as ErrorCode,
+    }
+
+    waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
+
+    const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+
+    expect(result).toBe(SafeCreationStatus.TIMEOUT)
+  })
+
+  it('returns SUCCESS if transaction was replaced', async () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TRANSACTION_REPLACED',
+      reason: 'repriced',
+    }
+    waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
+
+    const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+
+    expect(result).toBe(SafeCreationStatus.SUCCESS)
+  })
+
+  it('returns ERROR if transaction was cancelled', async () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TRANSACTION_REPLACED',
+      reason: 'cancelled',
+    }
+    waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
+
+    const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+
+    expect(result).toBe(SafeCreationStatus.ERROR)
+  })
+})
+
+describe('handleSafeCreationError', () => {
+  it('returns WALLET_REJECTED if the tx was rejected in the wallet', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'ACTION_REJECTED' as ErrorCode,
+      reason: '' as EthersTxReplacedReason,
+      receipt: {} as TransactionReceipt,
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.WALLET_REJECTED)
+  })
+
+  it('returns WALLET_REJECTED if the tx was rejected via WC', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'UNKNOWN_ERROR' as ErrorCode,
+      reason: '' as EthersTxReplacedReason,
+      receipt: {} as TransactionReceipt,
+      message: 'rejected',
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.WALLET_REJECTED)
+  })
+
+  it('returns ERROR if the tx was cancelled', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TRANSACTION_REPLACED' as ErrorCode,
+      reason: EthersTxReplacedReason.cancelled,
+      receipt: {} as TransactionReceipt,
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.ERROR)
+  })
+
+  it('returns SUCCESS if the tx was replaced', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TRANSACTION_REPLACED' as ErrorCode,
+      reason: EthersTxReplacedReason.replaced,
+      receipt: {} as TransactionReceipt,
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.SUCCESS)
+  })
+
+  it('returns SUCCESS if the tx was repriced', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'TRANSACTION_REPLACED' as ErrorCode,
+      reason: EthersTxReplacedReason.repriced,
+      receipt: {} as TransactionReceipt,
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.SUCCESS)
+  })
+
+  it('returns ERROR if the tx was not rejected, cancelled or replaced', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'UNKNOWN_ERROR' as ErrorCode,
+      reason: '' as EthersTxReplacedReason,
+      receipt: {} as TransactionReceipt,
+    }
+
+    const result = handleSafeCreationError(mockEthersError)
+
+    expect(result).toEqual(SafeCreationStatus.ERROR)
+  })
+
+  it('returns REVERTED if the tx failed', () => {
+    const mockEthersError = {
+      ...new Error(),
+      code: 'UNKNOWN_ERROR' as ErrorCode,
+      reason: '' as EthersTxReplacedReason,
+      receipt: {
         status: 0,
-      } as TransactionReceipt
+      } as TransactionReceipt,
+    }
 
-      waitForTxSpy.mockImplementationOnce(() => Promise.resolve(receipt))
+    const result = handleSafeCreationError(mockEthersError)
 
-      const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
+    expect(result).toEqual(SafeCreationStatus.REVERTED)
+  })
+})
 
-      expect(result).toBe(SafeCreationStatus.REVERTED)
-    })
+describe('createNewSafeViaRelayer', () => {
+  const owner1 = toBeHex('0x1', 20)
+  const owner2 = toBeHex('0x2', 20)
 
-    it('returns TIMEOUT if transaction could not be found within the timeout limit', async () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TIMEOUT' as ErrorCode,
-      }
+  const mockChainInfo = {
+    chainId: '5',
+    l2: false,
+  } as ChainInfo
 
-      waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
-
-      const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
-
-      expect(result).toBe(SafeCreationStatus.TIMEOUT)
-    })
-
-    it('returns SUCCESS if transaction was replaced', async () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TRANSACTION_REPLACED',
-        reason: 'repriced',
-      }
-      waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
-
-      const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
-
-      expect(result).toBe(SafeCreationStatus.SUCCESS)
-    })
-
-    it('returns ERROR if transaction was cancelled', async () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TRANSACTION_REPLACED',
-        reason: 'cancelled',
-      }
-      waitForTxSpy.mockImplementationOnce(() => Promise.reject(mockEthersError))
-
-      const result = await checkSafeCreationTx(provider, mockPendingTx, '0x0', jest.fn())
-
-      expect(result).toBe(SafeCreationStatus.ERROR)
-    })
+  beforeAll(() => {
+    jest.resetAllMocks()
+    jest.spyOn(web3, 'getWeb3ReadOnly').mockImplementation(() => provider)
   })
 
-  describe('handleSafeCreationError', () => {
-    it('returns WALLET_REJECTED if the tx was rejected in the wallet', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'ACTION_REJECTED' as ErrorCode,
-        reason: '' as EthersTxReplacedReason,
-        receipt: {} as TransactionReceipt,
-      }
+  it('returns taskId if create Safe successfully relayed', async () => {
+    const sponsoredCallSpy = jest.spyOn(relaying, 'sponsoredCall').mockResolvedValue({ taskId: '0x123' })
 
-      const result = handleSafeCreationError(mockEthersError)
+    const expectedSaltNonce = 69
+    const expectedThreshold = 1
+    const proxyFactoryAddress = await (await getReadOnlyProxyFactoryContract('5', LATEST_SAFE_VERSION)).getAddress()
+    const readOnlyFallbackHandlerContract = await getReadOnlyFallbackHandlerContract('5', LATEST_SAFE_VERSION)
+    const safeContractAddress = await (await getReadOnlyGnosisSafeContract(mockChainInfo)).getAddress()
 
-      expect(result).toEqual(SafeCreationStatus.WALLET_REJECTED)
-    })
+    const expectedInitializer = Gnosis_safe__factory.createInterface().encodeFunctionData('setup', [
+      [owner1, owner2],
+      expectedThreshold,
+      ZERO_ADDRESS,
+      EMPTY_DATA,
+      await readOnlyFallbackHandlerContract.getAddress(),
+      ZERO_ADDRESS,
+      0,
+      ZERO_ADDRESS,
+    ])
 
-    it('returns WALLET_REJECTED if the tx was rejected via WC', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'UNKNOWN_ERROR' as ErrorCode,
-        reason: '' as EthersTxReplacedReason,
-        receipt: {} as TransactionReceipt,
-        message: 'rejected',
-      }
+    const expectedCallData = Proxy_factory__factory.createInterface().encodeFunctionData('createProxyWithNonce', [
+      safeContractAddress,
+      expectedInitializer,
+      expectedSaltNonce,
+    ])
 
-      const result = handleSafeCreationError(mockEthersError)
+    const taskId = await relaySafeCreation(mockChainInfo, [owner1, owner2], expectedThreshold, expectedSaltNonce)
 
-      expect(result).toEqual(SafeCreationStatus.WALLET_REJECTED)
-    })
-
-    it('returns ERROR if the tx was cancelled', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TRANSACTION_REPLACED' as ErrorCode,
-        reason: EthersTxReplacedReason.cancelled,
-        receipt: {} as TransactionReceipt,
-      }
-
-      const result = handleSafeCreationError(mockEthersError)
-
-      expect(result).toEqual(SafeCreationStatus.ERROR)
-    })
-
-    it('returns SUCCESS if the tx was replaced', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TRANSACTION_REPLACED' as ErrorCode,
-        reason: EthersTxReplacedReason.replaced,
-        receipt: {} as TransactionReceipt,
-      }
-
-      const result = handleSafeCreationError(mockEthersError)
-
-      expect(result).toEqual(SafeCreationStatus.SUCCESS)
-    })
-
-    it('returns SUCCESS if the tx was repriced', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'TRANSACTION_REPLACED' as ErrorCode,
-        reason: EthersTxReplacedReason.repriced,
-        receipt: {} as TransactionReceipt,
-      }
-
-      const result = handleSafeCreationError(mockEthersError)
-
-      expect(result).toEqual(SafeCreationStatus.SUCCESS)
-    })
-
-    it('returns ERROR if the tx was not rejected, cancelled or replaced', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'UNKNOWN_ERROR' as ErrorCode,
-        reason: '' as EthersTxReplacedReason,
-        receipt: {} as TransactionReceipt,
-      }
-
-      const result = handleSafeCreationError(mockEthersError)
-
-      expect(result).toEqual(SafeCreationStatus.ERROR)
-    })
-
-    it('returns REVERTED if the tx failed', () => {
-      const mockEthersError = {
-        ...new Error(),
-        code: 'UNKNOWN_ERROR' as ErrorCode,
-        reason: '' as EthersTxReplacedReason,
-        receipt: {
-          status: 0,
-        } as TransactionReceipt,
-      }
-
-      const result = handleSafeCreationError(mockEthersError)
-
-      expect(result).toEqual(SafeCreationStatus.REVERTED)
-    })
-  })
-
-  describe('createNewSafeViaRelayer', () => {
-    const owner1 = toBeHex('0x1', 20)
-    const owner2 = toBeHex('0x2', 20)
-
-    const mockChainInfo = {
+    expect(taskId).toEqual('0x123')
+    expect(sponsoredCallSpy).toHaveBeenCalledTimes(1)
+    expect(sponsoredCallSpy).toHaveBeenCalledWith({
       chainId: '5',
-      l2: false,
-    } as ChainInfo
-
-    beforeAll(() => {
-      jest.resetAllMocks()
-      jest.spyOn(web3, 'getWeb3ReadOnly').mockImplementation(() => provider)
+      to: proxyFactoryAddress,
+      data: expectedCallData,
     })
+  })
 
-    it('returns taskId if create Safe successfully relayed', async () => {
-      const sponsoredCallSpy = jest.spyOn(relaying, 'sponsoredCall').mockResolvedValue({ taskId: '0x123' })
+  it('should throw an error if relaying fails', () => {
+    const relayFailedError = new Error('Relay failed')
 
-      const expectedSaltNonce = 69
-      const expectedThreshold = 1
-      const proxyFactoryAddress = await (await getReadOnlyProxyFactoryContract('5', LATEST_SAFE_VERSION)).getAddress()
-      const readOnlyFallbackHandlerContract = await getReadOnlyFallbackHandlerContract('5', LATEST_SAFE_VERSION)
-      const safeContractAddress = await (await getReadOnlyGnosisSafeContract(mockChainInfo)).getAddress()
+    jest.spyOn(relaying, 'sponsoredCall').mockRejectedValue(relayFailedError)
 
-      const expectedInitializer = Gnosis_safe__factory.createInterface().encodeFunctionData('setup', [
-        [owner1, owner2],
-        expectedThreshold,
-        ZERO_ADDRESS,
-        EMPTY_DATA,
-        await readOnlyFallbackHandlerContract.getAddress(),
-        ZERO_ADDRESS,
-        0,
-        ZERO_ADDRESS,
-      ])
-
-      const expectedCallData = Proxy_factory__factory.createInterface().encodeFunctionData('createProxyWithNonce', [
-        safeContractAddress,
-        expectedInitializer,
-        expectedSaltNonce,
-      ])
-
-      const taskId = await relaySafeCreation(mockChainInfo, [owner1, owner2], expectedThreshold, expectedSaltNonce)
-
-      expect(taskId).toEqual('0x123')
-      expect(sponsoredCallSpy).toHaveBeenCalledTimes(1)
-      expect(sponsoredCallSpy).toHaveBeenCalledWith({
-        chainId: '5',
-        to: proxyFactoryAddress,
-        data: expectedCallData,
-      })
-    })
-
-    it('should throw an error if relaying fails', () => {
-      const relayFailedError = new Error('Relay failed')
-
-      jest.spyOn(relaying, 'sponsoredCall').mockRejectedValue(relayFailedError)
-
-      expect(relaySafeCreation(mockChainInfo, [owner1, owner2], 1, 69)).rejects.toEqual(relayFailedError)
-    })
+    expect(relaySafeCreation(mockChainInfo, [owner1, owner2], 1, 69)).rejects.toEqual(relayFailedError)
   })
 })
