+++ src/features/counterfactual/utils.ts
@@ -1,7 +1,14 @@
+import type { NewSafeFormData } from '@/components/new-safe/create'
 import { LATEST_SAFE_VERSION } from '@/config/constants'
+import { AppRoutes } from '@/config/routes'
+import { addUndeployedSafe } from '@/features/counterfactual/store/undeployedSafeSlice'
+import type { AppDispatch } from '@/store'
+import { addOrUpdateSafe } from '@/store/addedSafesSlice'
+import { upsertAddressBookEntry } from '@/store/addressBookSlice'
 import { defaultSafeInfo } from '@/store/safeInfoSlice'
-import type { PredictedSafeProps } from '@safe-global/protocol-kit'
+import type { DeploySafeProps, PredictedSafeProps } from '@safe-global/protocol-kit'
 import { ZERO_ADDRESS } from '@safe-global/protocol-kit/dist/src/utils/constants'
+import type { SafeVersion } from '@safe-global/safe-core-sdk-types'
 import {
   type ChainInfo,
   ImplementationVersionState,
@@ -9,6 +16,7 @@ import {
   TokenType,
 } from '@safe-global/safe-gateway-typescript-sdk'
 import type { BrowserProvider } from 'ethers'
+import type { NextRouter } from 'next/router'
 
 export const getUndeployedSafeInfo = (undeployedSafe: PredictedSafeProps, address: string, chainId: string) => {
   return Promise.resolve({
@@ -45,4 +53,44 @@ export const getCounterfactualBalance = async (safeAddress: string, provider?: B
       },
     ],
   }
+}
+
+export const createCounterfactualSafe = (
+  chain: ChainInfo,
+  safeAddress: string,
+  saltNonce: string,
+  data: NewSafeFormData,
+  dispatch: AppDispatch,
+  props: DeploySafeProps,
+  router: NextRouter,
+) => {
+  const undeployedSafe = {
+    chainId: chain.chainId,
+    address: safeAddress,
+    safeProps: {
+      safeAccountConfig: props.safeAccountConfig,
+      safeDeploymentConfig: {
+        saltNonce,
+        safeVersion: LATEST_SAFE_VERSION as SafeVersion,
+      },
+    },
+  }
+
+  dispatch(addUndeployedSafe(undeployedSafe))
+  dispatch(upsertAddressBookEntry({ chainId: chain.chainId, address: safeAddress, name: data.name }))
+  dispatch(
+    addOrUpdateSafe({
+      safe: {
+        ...defaultSafeInfo,
+        address: { value: safeAddress, name: data.name },
+        threshold: data.threshold,
+        owners: data.owners.map((owner) => ({
+          value: owner.address,
+          name: owner.name || owner.ens,
+        })),
+        chainId: chain.chainId,
+      },
+    }),
+  )
+  router.push({ pathname: AppRoutes.home, query: { safe: `${chain.shortName}:${safeAddress}` } })
 }
